Move parse_timerange to TimeRange class
